# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * OutLawZ Labs Portfolio API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export type ProjectStatus = typeof ProjectStatus[keyof typeof ProjectStatus];


export const ProjectStatus = {
  active: 'active',
  archived: 'archived',
  wip: 'wip',
} as const;

export interface Project {
  id: number;
  title: string;
  slug: string;
  description: string;
  /** @nullable */
  long_description?: string | null;
  /** @nullable */
  problem?: string | null;
  /** @nullable */
  solution?: string | null;
  tech_stack?: string[];
  /** @nullable */
  category?: string | null;
  status: ProjectStatus;
  /** @nullable */
  cover_image?: string | null;
  gallery?: string[];
  /** @nullable */
  github_url?: string | null;
  /** @nullable */
  live_url?: string | null;
  featured: boolean;
  created_at: string;
}

export type ProjectInputStatus = typeof ProjectInputStatus[keyof typeof ProjectInputStatus];


export const ProjectInputStatus = {
  active: 'active',
  archived: 'archived',
  wip: 'wip',
} as const;

export interface ProjectInput {
  title: string;
  slug: string;
  description: string;
  long_description?: string;
  problem?: string;
  solution?: string;
  tech_stack?: string[];
  category?: string;
  status: ProjectInputStatus;
  cover_image?: string;
  gallery?: string[];
  github_url?: string;
  live_url?: string;
  featured?: boolean;
}

export type ProjectUpdateStatus = typeof ProjectUpdateStatus[keyof typeof ProjectUpdateStatus];


export const ProjectUpdateStatus = {
  active: 'active',
  archived: 'archived',
  wip: 'wip',
} as const;

export interface ProjectUpdate {
  title?: string;
  slug?: string;
  description?: string;
  long_description?: string;
  problem?: string;
  solution?: string;
  tech_stack?: string[];
  category?: string;
  status?: ProjectUpdateStatus;
  cover_image?: string;
  gallery?: string[];
  github_url?: string;
  live_url?: string;
  featured?: boolean;
}

export interface ContactMessage {
  id: number;
  name: string;
  email: string;
  message: string;
  read?: boolean;
  created_at: string;
}

export interface ContactInput {
  name: string;
  email: string;
  message: string;
}

export interface AdminLoginInput {
  password: string;
}

export interface AdminLoginResult {
  success: boolean;
  message?: string;
}

export interface AdminSession {
  authenticated: boolean;
}

export interface CategoryCount {
  category: string;
  count: number;
}

export interface PortfolioStats {
  totalProjects: number;
  featuredProjects: number;
  totalMessages: number;
  unreadMessages: number;
  projectsByCategory: CategoryCount[];
}

export type ListProjectsParams = {
category?: string;
search?: string;
featured?: string;
};