# lib/api-zod/src/generated/types/contactMessage.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * OutLawZ Labs Portfolio API
 * OpenAPI spec version: 0.1.0
 */

export interface ContactMessage {
  id: number;
  name: string;
  email: string;
  message: string;
  read?: boolean;
  created_at: string;
}