# lib/api-zod/src/generated/types/portfolioStats.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * OutLawZ Labs Portfolio API
 * OpenAPI spec version: 0.1.0
 */
import type { CategoryCount } from './categoryCount';

export interface PortfolioStats {
  totalProjects: number;
  featuredProjects: number;
  totalMessages: number;
  unreadMessages: number;
  projectsByCategory: CategoryCount[];
}