# lib/api-zod/src/generated/types/project.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * OutLawZ Labs Portfolio API
 * OpenAPI spec version: 0.1.0
 */
import type { ProjectStatus } from './projectStatus';

export interface Project {
  id: number;
  title: string;
  slug: string;
  description: string;
  /** @nullable */
  long_description?: string | null;
  /** @nullable */
  problem?: string | null;
  /** @nullable */
  solution?: string | null;
  tech_stack?: string[];
  /** @nullable */
  category?: string | null;
  status: ProjectStatus;
  /** @nullable */
  cover_image?: string | null;
  gallery?: string[];
  /** @nullable */
  github_url?: string | null;
  /** @nullable */
  live_url?: string | null;
  featured: boolean;
  created_at: string;
}