# lib/api-zod/src/generated/types/projectInput.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * OutLawZ Labs Portfolio API
 * OpenAPI spec version: 0.1.0
 */
import type { ProjectInputStatus } from './projectInputStatus';

export interface ProjectInput {
  title: string;
  slug: string;
  description: string;
  long_description?: string;
  problem?: string;
  solution?: string;
  tech_stack?: string[];
  category?: string;
  status: ProjectInputStatus;
  cover_image?: string;
  gallery?: string[];
  github_url?: string;
  live_url?: string;
  featured?: boolean;
}